+++ jax/_src/numpy/ufuncs.py
@@ -25,7 +25,6 @@ import operator
 import numpy as np
 
 from jax._src import core
-from jax._src import deprecations
 from jax._src import dtypes
 from jax._src.api import jit
 from jax._src.custom_derivatives import custom_jvp
@@ -1132,9 +1131,6 @@ def heaviside(x1: ArrayLike, x2: ArrayLike, /) -> Array:
                 _where(lax.gt(x1, zero), _lax_const(x1, 1), x2))
 
 
-deprecations.register("jax-numpy-hypot-complex")
-
-
 @implements(np.hypot, module='numpy')
 @jit
 def hypot(x1: ArrayLike, x2: ArrayLike, /) -> Array:
@@ -1143,13 +1139,9 @@ def hypot(x1: ArrayLike, x2: ArrayLike, /) -> Array:
   # TODO(micky774): Promote to ValueError when deprecation is complete
   # (began 2024-4-14).
   if dtypes.issubdtype(x1.dtype, np.complexfloating):
-    deprecations.warn(
-      "jax-numpy-hypot-complex",
-      "Passing complex-valued inputs to hypot is deprecated and will raise a "
-      "ValueError in the future. Please convert to real values first, such as "
-      "by using jnp.real or jnp.imag to take the real or imaginary components "
-      "respectively.",
-      stacklevel=2)
+    raise ValueError(
+      "jnp.hypot is not well defined for complex-valued inputs. "
+      "Please convert to real values first, such as by using abs(x)")
   x1, x2 = lax.abs(x1), lax.abs(x2)
   idx_inf = lax.bitwise_or(isposinf(x1), isposinf(x2))
   x1, x2 = maximum(x1, x2), minimum(x1, x2)